+++ instrumentation/logback/logback-appender-1.0/javaagent/src/test/java/io/opentelemetry/instrumentation/logback/appender/v1_0/LogbackTest.java
@@ -18,11 +18,4 @@ class LogbackTest extends AbstractLogbackTest {
   protected InstrumentationExtension testing() {
     return testing;
   }
-
-  @Override
-  protected boolean expectEventName() {
-    // the javaagent tests do not enable the deprecated capture-event-name setting
-    // see https://github.com/open-telemetry/semantic-conventions/issues/2913
-    return false;
-  }
 }